% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSNOTFile.R
\name{buildSNOTFile}
\alias{buildSNOTFile}
\title{buildSNOTFile}
\usage{
buildSNOTFile(
  dataSNOT,
  theme,
  frequence,
  site,
  station,
  datatype,
  minDate,
  maxDate,
  repsortie
)
}
\arguments{
\item{dataSNOT}{dataframe of data to be integrated}

\item{theme}{string of theme code of the data (ex. "meteosol")}

\item{frequence}{string of frecency code (ex. "sh" or "infraj")}

\item{site}{string of the code site (ex. "lgt")}

\item{station}{string of the code station (ex. "bm1")}

\item{datatype}{string of datatype code (ex. "meteosol")}

\item{minDate}{string of begin date (ex. \%d/\%m/\%Y)}

\item{maxDate}{string of end date (ex. \%d/\%m/\%Y)}

\item{repsortie}{string of the path to export file (default "../3_SNOT_meteosol/")}
}
\value{
Data file with name and structure for data-snot.cnrs.fr. Ready to be integrated in data-snot.cnrs.fr
}
\description{
Build and write a file with SNO-Tourbières format. Ready to be integrated into data-snot.snot.cnrs.fr
}
