% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeTimeTrend.R
\name{csvToSimpleDygraph}
\alias{csvToSimpleDygraph}
\title{csvToSimpleDygraph}
\usage{
csvToSimpleDygraph(compileFile, variableName)
}
\arguments{
\item{compileFile}{csv file (with "timestamp" column date (\%Y-\%m-\%d \%H:\%M:\%S)}

\item{variableName}{String of variable code into compileFile (ex. "WTD")}
}
\value{
html time trend with dygraph from each variable
}
\description{
Simple dygraph visualization from csv timeseries
}
\examples{
compileFilePath <- system.file("extdata","compilation/Compile_lgt_bm1_2020.csv", package = "toolboxMeteosol")
dataCompile <- read.csv(compileFilePath,header=TRUE,sep=";")
csvToSimpleDygraph(dataCompile,"WTD")
}
