% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeTimeTrend.R
\name{csvToXts}
\alias{csvToXts}
\title{csvToXts}
\usage{
csvToXts(compileFile, variableName)
}
\arguments{
\item{compileFile}{csv file (specific format)}

\item{variableName}{String of variable code into compileFile (ex. "WTD")}
}
\value{
xts object to used with dygraph with all variable selected ()
}
\description{
convert csv to xts format to use dygraph function
}
\examples{
compileFilePath <- system.file("extdata","compilation/Compile_lgt_bm1_2020.csv", package = "toolboxMeteosol")
compileFile <- read.csv(compileFilePath,header=TRUE,sep=";")
variableName <- c("TS_1_1_1","TS_1_2_1")
csvToXts(compileFile,variableName)
}
