% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeTimeTrend.R
\name{dataToXts}
\alias{dataToXts}
\title{dataToXts}
\usage{
dataToXts(compileFile, variableName, timestampColumn, timestampFormat)
}
\arguments{
\item{compileFile}{data.frame (specific format, add some test ?)}

\item{variableName}{String of variable code into compileFile (ex. "WTD", or c("WTD","TS_1_1_1"))}

\item{timestampColumn}{String of the timestampColumn with year, month, day, hour month and second(ex. "timestamp")}

\item{timestampFormat}{String of the timestampColumn format with year, month, day, hour month and second(ex. "\%Y-\%m-\%d \%H:\%M:\%S")}
}
\value{
xts object to used with dygraph with all variable selected ()
}
\description{
convert csv to xts format to use dygraph function
}
\examples{
compileFilePath <- system.file("extdata",
"compilation/Compile_lgt_bm1_2020.csv",
package = "toolboxMeteosol")
compileFile <- read.csv(compileFilePath,header=TRUE,sep=";")
timestampColumn <- "timestamp"
timestampFormat <- "\%Y-\%m-\%d \%H:\%M:\%S" 
variableName <- c("TS_1_1_1","TS_1_2_1")
dataToXts(compileFile,variableName,timestampColumn,timestampFormat)
}
