% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityControl.R
\name{qualityControlBM1}
\alias{qualityControlBM1}
\title{qualityControlBM1}
\usage{
qualityControlBM1(data_brute)
}
\arguments{
\item{data_brute}{a data.table}
}
\value{
A data.table with correction
}
\description{
Quality control of data.table from a bm1 station
}
