% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRawFilesBM.R
\name{readRawFilesBM}
\alias{readRawFilesBM}
\title{readRawFilesBM}
\usage{
readRawFilesBM(rawFilePath)
}
\arguments{
\item{rawFilePath}{Complete file path of raw file (ex. "~/SI_SNOT/processMeteoSol/inst/extdata/FR_LGT_001_20_036_0000.csv"). Need to respect format presented \href{https://sourcesup.renater.fr/www/si-snot/3.1_Sce_meteosol.html}{here}}
}
\value{
data.table meteosol semi-hourly
}
\description{
Read file from CR1000X datalogger, convert variable name and UTC+1 date configuration. Used for biometeorological station of SNO-Tourbières
}
\examples{
rawFilePath <- system.file("extdata",
"FR_LGT_001_20_036_0000.csv",
package = "toolboxMeteosol")
readRawFilesBM(rawFilePath)
}
