% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilHeatFluxCalcul.R
\name{soilHeatFluxCalcul}
\alias{soilHeatFluxCalcul}
\title{soilHeatFluxCalcul}
\usage{
soilHeatFluxCalcul(
  data_brute,
  SWC_variable,
  G_variable,
  Gqf_variable,
  Esen_variable
)
}
\arguments{
\item{data_brute}{data.table}

\item{SWC_variable}{string vector of SWC variable}

\item{G_variable}{string vector of Gvariable}

\item{Gqf_variable}{string vector of Gqf variable}

\item{Esen_variable}{string vector of Esen variable}
}
\value{
data.table meteosol semi-hourly
}
\description{
Soil Heat Flux calculation
}
