% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilWaterConcentCalcul.R
\name{soilWaterConcentCalcul}
\alias{soilWaterConcentCalcul}
\title{soilWaterConcentCalcul}
\usage{
soilWaterConcentCalcul(data_brute, SWC_variable, Pe_variable)
}
\arguments{
\item{data_brute}{data.table with SWC and Pe variables}

\item{SWC_variable}{string vector of SWC variables (ex. SWC_X_X_X)}

\item{Pe_variable}{string vector of Pe variables (ex. Pe_X_X_X)}
}
\value{
data.table with SWC values
}
\description{
Soil Water Content calculation on a data.table with SWC and PE variable
}
