% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflowMeteosolData.R
\name{workflowMeteosolData}
\alias{workflowMeteosolData}
\title{Data processing of raw files from biometeorological station of SNO-Tourbières}
\usage{
workflowMeteosolData(
  repFile,
  repTraitement,
  repOut,
  yearsToIntegrate,
  Site,
  Station
)
}
\arguments{
\item{repFile}{repository of raw files (ex. 1_BRUTE_meteosol/03-02-2016_actuel/)}

\item{repTraitement}{out repository for temp files}

\item{repOut}{out repository for final files.  Ready to integrate to data-snot.cnrs.fr}

\item{yearsToIntegrate}{String of the years to integrate (yy) (ex. "19" for 2019...,"all" pour all years)}

\item{Site}{String of the code site code (ex. "lgt", "bdz", "ldm", or "frn")}

\item{Station}{String of the station code (ex. "bm1", "bm2", "bm3"...)}
}
\value{
A csv file ready to be integrated in data-snot.cnrs.fr
}
\description{
Compilation + qualitycontrol of raw files from 1_BRUTE_meteosol/
}
\examples{
workflowMeteosolData(repFile=system.file("extdata", package = "toolboxMeteosol"),
			repTraitement="~/",
			repOut="~/",
			yearsToIntegrate="20",
			Site="lgt",
			Station="bm1")
}
