% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilWaterConcentCalcul.R
\name{soilWaterConcentCalcul}
\alias{soilWaterConcentCalcul}
\title{soilWaterConcentCalcul}
\usage{
soilWaterConcentCalcul(data_brute, SWC_variable, Pe_variable)
}
\arguments{
\item{data_brute}{data.table}

\item{SWC_variable}{string vector of SWC variable}

\item{Pe_variable}{string vector of Pe variable}
}
\value{
Retourne une data.table meteosol semi-horaire
}
\description{
Contrôle qualité d'une data.table
}
